requirements.txt
arrow==1.3.0
asgiref==3.8.1
beautifulsoup4==4.12.3
binaryornot==0.4.4
certifi==2024.2.2
cfgv==3.4.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cookiecutter==2.5.0
distlib==0.3.8
Django==5.1
django-cors-headers==4.4.0
django-filter==24.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
filelock==3.13.1
identify==2.5.33
idna==3.6
Jinja2==3.1.3
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
nodeenv==1.8.0
pdfkit==1.0.0
platformdirs==4.2.0
pre-commit==3.6.0
Pygments==2.17.2
PyJWT==2.9.0
python-dateutil==2.8.2
python-slugify==8.0.3
PyYAML==6.0.1
requests==2.31.0
rich==13.7.0
setuptools==69.0.3
six==1.16.0
soupsieve==2.5
sqlparse==0.5.1
text-unidecode==1.3
types-python-dateutil==2.8.19.20240106
tzdata==2024.1
urllib3==2.2.0
virtualenv==20.25.0
